extracted queue register related into QueueRegistryService
